Re-use more code when testing
