+++ tests/fixtures/setup.ts
@@ -1,24 +1,6 @@
 import { afterEach, beforeEach } from 'bun:test';
 import path from 'node:path';
-import { Engine } from '@ronin/engine';
-import { BunDriver } from '@ronin/engine/drivers/bun';
-import { MemoryResolver } from '@ronin/engine/resolvers/memory';
-import type { Row, Statement } from '@ronin/engine/types';
-
-const engine = new Engine({
-  resolvers: [
-    new MemoryResolver({
-      driver: new BunDriver(),
-    }),
-  ],
-});
-
-export const queryDatabase = async (
-  statements: Array<Statement>,
-): Promise<Array<Array<Row>>> => {
-  const results = await engine.queryDatabase('test', statements);
-  return results.map((result) => result.rows);
-};
+import { engine, queryDatabase } from '@/fixtures/utils';
 
 beforeEach(async () => {
   await engine.createDatabase('test');
@@ -32,7 +14,7 @@ beforeEach(async () => {
     .map((statement) => statement.replaceAll('\n', ''));
   const statements = statementStrings.map((statement) => ({ statement, params: [] }));
 
-  await engine.queryDatabase('test', statements);
+  await queryDatabase(statements);
 });
 
 afterEach(async () => {
